% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_form.R
\name{create_text}
\alias{create_text}
\title{Create markdown text}
\usage{
create_text(ssblist, tag)
}
\arguments{
\item{ssblist}{The current list of methds functions}

\item{tag}{The process numbers (GSBPM one decimal) to identify}
}
\description{
Create markdown text
}
\author{
Susie Jentoft
}
