% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_form.R
\name{metodebib_input}
\alias{metodebib_input}
\title{Input form for adding a function to Metodebibliotek}
\usage{
metodebib_input()
}
\value{
the function and its details are added to the data file and README and tabel files are rendered
}
\description{
Input form for adding a function to Metodebibliotek
}
