% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateSSB.R, R/PxWebApiData.R, R/SSBtools.R,
%   R/SmallCountRounding.R, R/easySdcTable.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{AkuData}
\alias{ApiData1}
\alias{ApiData2}
\alias{ApiData12}
\alias{GetApiData1}
\alias{GetApiData2}
\alias{GetApiData12}
\alias{SSBtoolsData}
\alias{HierarchyCompute}
\alias{SmallCountData}
\alias{EasyData}
\alias{ProtectKostra}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CalibrateSSB}{\code{\link[CalibrateSSB]{AkuData}}}

  \item{Kostra}{\code{\link[Kostra]{ProtectKostra}}}

  \item{PxWebApiData}{\code{\link[PxWebApiData:ApiData]{ApiData1}}, \code{\link[PxWebApiData:ApiData]{ApiData12}}, \code{\link[PxWebApiData:ApiData]{ApiData2}}, \code{\link[PxWebApiData:ApiData]{GetApiData1}}, \code{\link[PxWebApiData:ApiData]{GetApiData12}}, \code{\link[PxWebApiData:ApiData]{GetApiData2}}}

  \item{SSBtools}{\code{\link[SSBtools]{HierarchyCompute}}, \code{\link[SSBtools]{SSBtoolsData}}}

  \item{SmallCountRounding}{\code{\link[SmallCountRounding]{SmallCountData}}}

  \item{easySdcTable}{\code{\link[easySdcTable]{EasyData}}}
}}

